src/package/pyproject.toml
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.


[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dataplexutils_metadata_wizard"
authors = [
    {name = "Jakub Skuratowicz", email = "kubas@google.com"}
]
description = "Google Dataplex metadata automation using Generative AI"
version = "0.0.2"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas==2.2.2",
    "toml==0.10.2",
    "google-cloud-bigquery==3.21.0",
    "google-cloud-datacatalog==3.19.0",
    "google-cloud-datacatalog-lineage==0.3.6",
    "google-cloud-dataplex==2.3.1",
    "google-cloud-aiplatform>=1.48.0"
]